XmlInterface
