+++ tests/Feature/EInvoice/Verifactu/VerifactuModelTest.php
@@ -1,5 +1,15 @@
 <?php
 
+/**
+ * Invoice Ninja (https://invoiceninja.com).
+ *
+ * @link https://github.com/invoiceninja/invoiceninja source repository
+ *
+ * @copyright Copyright (c) 2025. Invoice Ninja LLC (https://invoiceninja.com)
+ *
+ * @license https://www.elastic.co/licensing/elastic-license
+ */
+
 namespace Tests\Feature\EInvoice\Verifactu;
 
 use App\Services\EDocument\Standards\Verifactu\Models\Cupon;
